third_party/WebKit/LayoutTests/fast/dom/HTMLElement/set-value-caseinsensitive-expected.txt

When contentEditable is set to a new value, the new value can be case-insensitive.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("div1").getAttribute("contentEditable") is "true"
PASS document.getElementById("div1").contentEditable is "true"
PASS document.getElementById("div1").isContentEditable is true
FIXME: isContentEditable is not working properly. Related to https://bugs.webkit.org/show_bug.cgi?id=52058
PASS window.getComputedStyle(div1, "").getPropertyValue("-webkit-user-modify") is "read-write"
PASS document.getElementById("p1").getAttribute("contentEditable") is "false"
PASS document.getElementById("p1").contentEditable is "false"
PASS document.getElementById("p1").isContentEditable is false
PASS window.getComputedStyle(p1, "").getPropertyValue("-webkit-user-modify") is "read-only"
PASS document.getElementById("p2").hasAttribute("contentEditable") is false
PASS document.getElementById("p2").contentEditable is "inherit"
PASS document.getElementById("p2").isContentEditable is true
PASS window.getComputedStyle(p2, "").getPropertyValue("-webkit-user-modify") is "read-write"
PASS document.getElementById("div3").hasAttribute("contentEditable") is true
PASS document.getElementById("div3").contentEditable is "plaintext-only"
PASS document.getElementById("div3").isContentEditable is true
PASS window.getComputedStyle(div3, "").getPropertyValue("-webkit-user-modify") is "read-write-plaintext-only"
PASS document.getElementById("p3").hasAttribute("contentEditable") is false
PASS document.getElementById("p3").contentEditable is "inherit"
PASS document.getElementById("p3").isContentEditable is true
PASS window.getComputedStyle(p3, "").getPropertyValue("-webkit-user-modify") is "read-write-plaintext-only"
PASS successfullyParsed is true

TEST COMPLETE
